Move error handling to its own module and don't colorize if it already has ansi (#506)
