+++ package.json
@@ -17,6 +17,7 @@
     "esprima": "^4.0.1",
     "fauna": "^2.4.0",
     "faunadb": "^4.5.4",
+    "has-ansi": "^6.0.0",
     "inquirer": "^12.0.0",
     "luxon": "^3.5.0",
     "open": "10.1.0",
