+++ src/commands/query.mjs
@@ -2,13 +2,15 @@
 
 import { container } from "../cli.mjs";
 import {
-  CommandError,
-  isUnknownError,
   resolveFormat,
   validateDatabaseOrSecret,
-  ValidationError,
   yargsWithCommonConfigurableQueryOptions,
 } from "../lib/command-helpers.mjs";
+import {
+  CommandError,
+  isUnknownError,
+  ValidationError,
+} from "../lib/errors.mjs";
 import {
   formatError,
   formatQueryResponse,
@@ -83,10 +85,10 @@ async function queryCommand(argv) {
       timeout,
       typecheck,
       apiVersion,
-      color,
-      raw,
       performanceHints,
       summary,
+      color,
+      raw,
     } = argv;
 
     // If we're writing to a file, don't colorize the output regardless of the user's preference
@@ -123,7 +125,7 @@ async function queryCommand(argv) {
     if (argv.output) {
       container.resolve("fs").writeFileSync(argv.output, output);
     } else {
-      logger.stdout(output);
+      container.resolve("logger").stdout(output);
     }
 
     return results;
